Site search now searches for categories as well as posts.
To do this, the previous search function has been rewritten so that it will search a key that is passed as an argument. The search isn't perfect. Ideally, searching for something like “river rejuvenation” would produce the page titled “rejuvenation” in the river category but at the moment it won't.