Converted riffles & pools diagram to SVG format.
The SVG code for this is _horrible_ the program I use to make the diagram (Sketch) wouldn't export SVG code to rotate the arrows correctly so I had to export it to EPS and convert that to SVG with some dodgy online tool. Yuck. If I get chance, I'll try and manually fix the SVG code exported by Sketch.