Major scrolling animation changes
First up, animated scrolling is now much smoother.

Second, the scrolling animation on the homepage is now a quintic-out
animation. This makes the homepage feel much more responsive.

Third, the scrolling animation for footnotes is now an
exponential-in-out animation. It's less jarring when the animation
starts and stops but is faster when scrolling through lots of text.