Update 2012-01-04-tides.md
How much added or taken weight from the moon would cause a significant change in tides? For example if i wanted to mine water or helium 3 how many tons would I have to take to see the tidal patterns change or If we added mass to the moon by putting a colony or space station how much weight would be needed to increase the gravitational pull?